Change pakage name from com.framgia to com.sunasterisk
