+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.framgia.musixmatch"
+        applicationId "com.sunasterisk.musixmatch"
         minSdkVersion 22
         targetSdkVersion 28
         versionCode 1
