+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.framgia.musixmatch">
+    package="com.sunasterisk.musixmatch">
 
     <application
         android:name=".application.MusixMatchApplication"
@@ -12,8 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".ui.main.MainActivity" />
-        <activity android:name=".ui.splash.SplashActivity"
+        <activity android:name="ui.main.MainActivity" />
+        <activity
+            android:name="ui.splash.SplashActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,4 +25,3 @@
     </application>
 
 </manifest>
-
