--- app/src/main/java/com/framgia/musixmatch/application/MusixMatchApplication.java
@@ -1,17 +1,0 @@
-package com.framgia.musixmatch.application;
-
-import android.app.Application;
-
-public class MusixMatchApplication extends Application {
-    private static MusixMatchApplication sApplication;
-
-    public static MusixMatchApplication getInstance(){
-        return sApplication;
-    }
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        sApplication = this;
-    }
-}
