--- app/src/main/java/com/framgia/musixmatch/data/source/ArtistDataSource.java
@@ -1,11 +1,0 @@
-package com.framgia.musixmatch.data.source;
-
-import com.framgia.musixmatch.data.model.Artist;
-
-import java.util.List;
-
-public interface ArtistDataSource {
-    interface Local {
-        void getArtists(Callback<List<Artist>> callback);
-    }
-}
