--- app/src/main/java/com/framgia/musixmatch/data/source/Callback.java
@@ -1,7 +1,0 @@
-package com.framgia.musixmatch.data.source;
-
-public interface Callback<T> {
-    void getDataSuccess(T data);
-
-    void getDataFailure(Exception e);
-}
