--- app/src/main/java/com/framgia/musixmatch/ui/main/MainActivity.java
@@ -1,15 +1,0 @@
-package com.framgia.musixmatch.ui.main;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import com.framgia.musixmatch.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
