+++ app/src/main/java/com/sunasterisk/musixmatch/data/repository/AlbumRepository.java
@@ -1,0 +1,28 @@
+package com.sunasterisk.musixmatch.data.repository;
+
+import com.sunasterisk.musixmatch.data.model.Album;
+import com.sunasterisk.musixmatch.data.source.AlbumDataSource;
+import com.sunasterisk.musixmatch.data.source.Callback;
+
+import java.util.List;
+
+public class AlbumRepository implements AlbumDataSource.Local {
+    private static AlbumRepository sInstance;
+    private AlbumDataSource.Local mLocal;
+
+    private AlbumRepository(AlbumDataSource.Local local) {
+        mLocal = local;
+    }
+
+    public static AlbumRepository getInstance(AlbumDataSource.Local local) {
+        if (sInstance == null) {
+            sInstance = new AlbumRepository(local);
+        }
+        return sInstance;
+    }
+
+    @Override
+    public void getAlbums(Callback<List<Album>> callback) {
+        mLocal.getAlbums(callback);
+    }
+}
