+++ app/src/main/java/com/sunasterisk/musixmatch/data/source/AlbumDataSource.java
@@ -1,0 +1,11 @@
+package com.sunasterisk.musixmatch.data.source;
+
+import com.sunasterisk.musixmatch.data.model.Album;
+
+import java.util.List;
+
+public interface AlbumDataSource {
+    interface Local {
+        void getAlbums(Callback<List<Album>> callback);
+    }
+}
