+++ app/src/main/java/com/sunasterisk/musixmatch/data/source/Callback.java
@@ -1,0 +1,7 @@
+package com.sunasterisk.musixmatch.data.source;
+
+public interface Callback<T> {
+    void getDataSuccess(T data);
+
+    void getDataFailure(Exception e);
+}
