+++ app/src/main/java/com/sunasterisk/musixmatch/data/source/TrackDataSource.java
@@ -1,0 +1,11 @@
+package com.sunasterisk.musixmatch.data.source;
+
+import com.sunasterisk.musixmatch.data.model.Track;
+
+import java.util.List;
+
+public interface TrackDataSource {
+    interface Local {
+        void getTracks(Callback<List<Track>> callback);
+    }
+}
