+++ app/src/main/java/com/sunasterisk/musixmatch/ui/main/MainActivity.java
@@ -1,0 +1,15 @@
+package com.sunasterisk.musixmatch.ui.main;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.sunasterisk.musixmatch.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
