+++ app/src/main/res/layout/activity_playing.xml
@@ -85,7 +85,7 @@
         app:layout_constraintStart_toEndOf="@+id/button_back"
         app:layout_constraintTop_toTopOf="@+id/guideline3" />
 
-    <com.framgia.musixmatch.utils.widget.RepeatButtonView
+    <com.sunasterisk.musixmatch.utils.widget.RepeatButtonView
         android:id="@+id/button_repeat"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
